+++ src/components/cardEducation/cardEducation.module.css
@@ -1,0 +1,34 @@
+.content {
+  height: 128px;
+  width: 540px;
+  padding: 24px;
+}
+
+.item {
+  justify-content: start;
+  margin-bottom: 20px;
+}
+
+.title {
+  margin: 8px 0 0 0;
+  padding: 0;
+  font-size: 20px;
+  font-weight: 500;
+  line-height: 24px;
+}
+
+.company {
+  margin: 4px 0 20px 0;
+  padding: 0;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 20px;
+}
+
+.time {
+  margin: 0;
+  padding: 0;
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 24px;
+}
